project/requirements.txt
pandas==1.4.3
pyfiglet==0.8.post1
PyJWT==2.6.0
requests==2.28.1
termcolor==1.1.0
